+++ package.json
@@ -9,7 +9,7 @@
     "start": "next start",
     "lint": "next lint",
     "prisma.studio": "npx prisma studio",
-    "seed": "node -r dotenv/config ./scripts/seed.js",
+    "seed": "node --loader ts-node/esm prisma/seed.ts",
     "prisma.migrate": "npx prisma migrate dev",
     "prisma.push": "npx prisma db push",
     "postinstall": "prisma generate"
@@ -35,8 +35,5 @@
     "tailwindcss": "^3.4.1",
     "ts-node": "^10.9.2",
     "typescript": "^5.5.4"
-  },
-  "prisma": {
-    "seed": "ts-node prisma/seed.js"
   }
 }
