+++ src/app/specialists/page.tsx
@@ -1,5 +1,6 @@
 // import { createClient } from "@/utils/supabase/server";
-import { createClient } from '@supabase/supabase-js'
+import prisma from "@/lib/prismaClient";
+import { createClient } from "@supabase/supabase-js";
 
 interface Specialist {
   id: string;
@@ -7,14 +8,34 @@ interface Specialist {
   lastName: string;
 }
 interface SpecialistsPageProps {
-  specialists: Specialist[]
+  specialists: Specialist[];
 }
 
-const SpecialistsPage:React.FC<SpecialistsPageProps> = ({ specialists }) => {
-  console.log('Specialists in component:', specialists);
+export async function fetchSpecialists(): Promise<Specialist[]> {
+  try {
+    const specialists = await prisma.specialist.findMany({
+      select: {
+        id: true,
+        firstName: true,
+        lastName: true,
+      },
+    });
+    return specialists;
+  } catch (error) {
+    console.error("Error fetching specialists:", error);
+    return [];
+  }
+}
+interface SpecialistsPageProps {
+  specialists: Specialist[];
+}
+
+const SpecialistsPage: React.FC<SpecialistsPageProps> = ({ specialists }) => {
   return (
     <main className="text-center pt-32 px-5">
-      <h1 className="text-4xl md:text-5xl font-bold mb-5">Wszyscy specjaliści</h1>
+      <h1 className="text-4xl md:text-5xl font-bold mb-5">
+        Wszyscy specjaliści
+      </h1>
       <ul>
         {specialists.map((specialist) => (
           <li key={specialist.id}>
@@ -24,28 +45,11 @@ const SpecialistsPage:React.FC<SpecialistsPageProps> = ({ specialists }) => {
       </ul>
     </main>
   );
-}
- async function fetchSpecialists() {
-  const supabase = createClient(
-    process.env.NEXT_PUBLIC_SUPABASE_URL!,
-    process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY!
-  )
-  // const supabase = createClient();
-  const { data: specialists, error } = await supabase
-    .from('specialists')
-    .select();
+};
 
-  if (error) {
-    console.error('Error fetching specialists:', error);
-    return [];
-  }
-  console.log('Fetched specialists:', specialists);
-  return specialists as Specialist[]
-}
-
-export default async function Page() {
-  
+const Page = async () => {
   const specialists = await fetchSpecialists();
-  console.log('Specialists in Page:', specialists);
   return <SpecialistsPage specialists={specialists} />;
-}
+};
+
+export default Page;
