+++ tsconfig.json
@@ -21,6 +21,6 @@
       "@/*": ["./src/*"]
     }
   },
-  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts", "src/data/seed.ts", "scripts/seed.js", "prisma/seed.js"],
+  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts", "src/data/seed.ts", "scripts/seed.js", "prisma/**/*.ts", "prisma/seed.ts"],
   "exclude": ["node_modules"]
 }
